+++ src/main/java/lotto/domain/manager/LottoManager.java
@@ -5,7 +5,7 @@ import lotto.domain.lotto.LottoRepository;
 import lotto.domain.lottomachine.LottoMachine;
 import lotto.domain.rating.Rating;
 import lotto.domain.rating.RatingInfo;
-import lotto.domain.statistics.WinningLotto;
+import lotto.domain.rating.WinningLotto;
 
 public class LottoManager {
